Random recipes and ingredients page added
Added a random recipes part in the landing page where we make a few api
calls o fetch random recipes.

Added Recipe Component.

Added Ingredient components which are in the Recipe Component.

Ingredients components in Recipes are links that link to their own page
where we display information about the ingredient.

Ingredient pages now have a section that showcases the recipes that
feature the said ingredient.

Due to the way the api works the above recipes do not show their
ingredients like the random ones in the landing page. This will require
seperate fetches to the recipes by their names or id's for each and
every one of them.